fix(activemodel): distinguish shared and isolated namespaces in ModelName
Rails derives `param_key` / `route_key` from `@unnamespaced` only when
`Name.new` receives a `namespace` argument (naming.rb:171, :180-182), and
`model_name` passes one only for a module answering `use_relative_model_naming?`
(naming.rb:271-276). trails' `ModelName` always produced the prefix-dropped
shape, so a namespaced model could not spell Rails' shared-namespace output.

The namespace object form now carries `useRelativeModelNaming`, which selects
the isolated shape; a string or segment array is the shared one. `naming.test.ts`
asserts Rails' literal values throughout: `NamingTest` route/param keys are
`post_track_backs` / `post_track_back`, and
`NamingWithNamespacedModelInSharedNamespaceTest` asserts `blog_posts` /
`blog_post`.

Also splits the three TS-only NestedError tests into
`nested-error.trails.test.ts`, leaving `nested-error.test.ts` a straight port of
`nested_error_test.rb`.

Lowers the activemodel assertion-value mark from 69 to 65.
