+++ scripts/test-compare/assertion-mismatch-mark.json
@@ -23,7 +23,7 @@
     "activemodel": {
       "assertionCount": 456,
       "kind": 687,
-      "value": 69
+      "value": 65
     },
     "activerecord": {
       "assertionCount": 1977,
